extract version-stage helpers out of agent-configurations utils
Move VersionStage enum (was AgentConfigurationBasicStage), pickHighestStage,
isProdTag, isBasicStage, sortTags into a new utils/version-stages.ts. The
prompt-version UI (PromptTab, PromptCard, LoadedPromptDisplay, PromptLibraryMenu,
StageTag, VersionTagList, AgentRunnerPromptCard) now imports from there.

utils/agent-configurations.ts keeps only formatBlueprintValue. Dead exports
(AGENT_CONFIGURATION_METADATA_KEY, AGENT_CONFIGURATION_PROD_ENV_NAME, isStageTag)
removed since nothing else references them.
